% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fur.R
\name{fur}
\alias{fur}
\title{FUR}
\usage{
fur(input_rkgs, subit_len_list, search_radius, starting_rkg = c())
}
\arguments{
\item{input_rkgs}{a \code{n} by \code{k} matrix of \code{k} rankings of \code{n}
objects, where each row is a complete ranking.}

\item{subit_len_list}{a vector containing positive integer(s) for the subiteration
lengths to \emph{Subiterative Convergence}. Suggested values are between 2 and 8, and
smaller subiteration lengths will result in less run-time.}

\item{search_radius}{a positive integer for the maximum amount allowed to increase
or decrease the rank of each object in the \emph{Greedy Algorithm}. The default value
of \code{0} is used to update the rank of each object to all possible ranks. It is
recommended to use a search radius of less than or equal to 30.}

\item{starting_rkg}{a vector of length n with an initial ranking to begin FUR on. If
the default value of an empty vector is used, then the mean seed ranking is calculated
as initial input to FUR.}
}
\value{
A matrix containing the consensus ranking, total Kemeny distance, average
tau correlation coefficient.
}
\description{
\emph{FUR} contains three branches - Fixed, Update, and Range - to reach
a consensus ranking by iteratively using \emph{Subiterative Convergence} and \emph{Greedy Algorithm}.
See 'Details' for more information on each branch.
}
\details{
The Fixed branch applies \emph{Subiterative Convergence} using only one subiteration
length at at time.

The Update branch runs \emph{Subiterative Convergence} using the first
subiteration length in the subiteration length list, and then uses the output in the
next call to \emph{Subiterative Convergence} with the next subiteration length in the list.
This process repeats until every subiteration length in the list has been used.

The Range branch calls \emph{Subiterative Convergence} on all subiteration lengths in the list
and only retains the best ranking among these separate calls.

The output from the \emph{Subiterative Convergence} calls are fed into the \emph{Greedy Algorithm}
as its seed ranking, and the FUR algorithm is terminated when the input to the
\emph{Greedy Algorithm} converges to the output and all branches have been executed at
least once.
}
\examples{
## One subiteration length
input_rkgs <- matrix(c(3, 2, 5, 1, 2, 3, 1, 2, 5, 1, 3, 4, 4, 5, 4, 5, 1, 4, 2, 3), ncol = 5)
subit_len_list <- 2
search_radius <- 1
fur(input_rkgs, subit_len_list, search_radius) # Determined the consensus ranking, total Kemeny
                                              # distance, and average tau correlation coefficient

## Multiple subiteration lengths
input_rkgs <- matrix(c(3, 2, 5, 1, 2, 3, 1, 2, 5, 1, 3, 4, 4, 5, 4, 5, 1, 4, 2, 3), ncol = 5)
subit_len_list <- c(2,3)
search_radius <- 1
fur(input_rkgs, subit_len_list, search_radius)

## Included dataset of 15 input rankings of 50 objects
data(data50x15)
input_rkgs <- t(as.matrix(data50x15[, -1]))
subit_len_list <- c(2, 3)
search_radius <- 1
fur(input_rkgs, subit_len_list, search_radius)

}
\references{
Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative
convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210.
\url{https://doi.org/10.1016/j.cor.2018.06.007}
}
\seealso{
\code{\link{subit_convergence}}, \code{\link{rap_greedy_alg}}
}
