% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_rkgs.R
\name{get_sub_rkgs}
\alias{get_sub_rkgs}
\title{Determine a Subset of Rankings for Greedy Algorithm}
\usage{
get_sub_rkgs(seed_rkg, move_ind, search_radius)
}
\arguments{
\item{seed_rkg}{a vector containing the initial ranking on which subsequent moves will be based.}

\item{move_ind}{an integer representing the index of the object to be moved.}

\item{search_radius}{a positive integer of the maximum shift allowed to increase or decrease
the rank of the object.}
}
\value{
A matrix of the subset of rankings, with each row as a new ranking with the moved object.
}
\description{
This function creates the new rankings for \emph{Greedy Algorithm}
after moving an object to a higher and lower rank within the search radius.
Cyclical moving is permitted in the event of overflow.
}
\seealso{
\code{\link{rap_greedy_alg}}
}
\keyword{internal}
