% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kemeny_int.R
\name{mod_kemeny_int}
\alias{mod_kemeny_int}
\title{Modified Kemeny (Hidden from User)}
\usage{
mod_kemeny_int(input_rkgs, universe_rkgs, obj_pairs)
}
\arguments{
\item{input_rkgs}{a \code{n} by \code{k} matrix of \code{k} rankings of \code{n}
objects, where each row is a complete ranking.}

\item{universe_rkgs}{a matrix containing all possible permutations of ranking
n objects. Each row in this matrix represents one permuted ranking.}

\item{obj_pairs}{a \code{2} by \code{n choose 2} matrix of all combinations of
object pairs of n objects, where each column contains a pair of object indices.}
}
\value{
A matrix containing the consensus ranking, total Kemeny distance, and
average tau correlation coefficient.
}
\description{
\emph{Modified Kemeny} algorithm determines the consensus ranking using
the set of all possible rankings compared to the input rankings. This algorithm
based on minimizing the combined Kemeny distance compared to the input rankings;
however, see 'Details' for additional criteria to determine the consensus
ranking if multiple rankings have the minimum total Kemeny distance.
}
\details{
If multiple universe rankings have the minimum total Kemeny distance,
then the smallest maximum Kemeny distance is considered. If there are multiple
rankings remaining, the ranking(s) with the smallest variance are kept. Finally,
if multiple rankings exist after filtering these conditions, then the first
ranking is chosen as the single consensus ranking.
}
