% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_all_totalK.R
\name{score_all_totalK}
\alias{score_all_totalK}
\title{Score All Rankings versus Input Rankings}
\usage{
score_all_totalK(rkgs, input_rkgs, pairs)
}
\arguments{
\item{rkgs}{a matrix of rankings to be compared against the input rankings. Each row must be a
complete ranking, meaning that all of the objects have a rank.}

\item{input_rkgs}{a \code{n} by \code{k} matrix of \code{k} rankings of \code{n}
objects, where each row is a complete ranking.}

\item{pairs}{a \code{2} by \code{n choose 2} matrix of all combinations of
object pairs of n objects, where each column contains a pair of object indices.}
}
\value{
A vector of the total Kemeny distance, the maximum Kemeny distance of
the individual Kemeny distances, and the variance of the individual Kemeny distances.
}
\description{
Calculates the total Kemeny distance for a set of rankings versus the
input rankings. A ranking and one input ranking has an individual total Kemeny
distance, and thus, the total Kemeny distance is the sum of the \code{k}
Kemeny distances between the ranking and each kemeny distance. This function is
used in \emph{Modified Kemeny}, \emph{Subiterative Convergence},
\emph{Greedy Algorithm}, and \emph{Seed-Based Iteration}.
}
\seealso{
\code{\link{mod_kemeny}}, \code{\link{subit_convergence}}, \code{\link{rap_greedy_alg}},
\code{\link{seed_based_iteration}}
}
