% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_one_totalK.R
\name{score_one_totalK}
\alias{score_one_totalK}
\title{Score One Ranking versus Input Rankings}
\usage{
score_one_totalK(rkg, input_rkgs, pairs)
}
\arguments{
\item{rkg}{a vector to be compared against the input rankings. The ranking
should be complete which means that all of the objects have a rank.}

\item{input_rkgs}{a \code{n} by \code{k} matrix of \code{k} rankings of \code{n}
objects, where each row is a complete ranking.}

\item{pairs}{a \code{2} by \code{n choose 2} matrix of all combinations of
object pairs of n objects, where each column contains a pair of object indices.}
}
\value{
A vector of the total Kemeny distance, the maximum Kemeny distance of
the individual Kemeny distances, and the variance of the individual Kemeny distances.
}
\description{
Calculates the total Kemeny distance for one ranking versus the
input rankings. A ranking and one input ranking has an individual total Kemeny
distance, and thus, the total Kemeny distance is the sum of the \code{k}
Kemeny distances between the ranking and each kemeny distance.
}
\details{
This function is eventually used to get the total Kemeny
distances for a set of rankings versus the input rankings, and it is also
used in \emph{Subiterative Convergence}, \emph{Greedy Algorithm}, and \emph{FUR} to determine the
total Kemeny distance for the seed ranking or output ranking.
}
\seealso{
\code{\link{score_all_totalK}}, \code{\link{subit_convergence}}, \code{\link{rap_greedy_alg}},
\code{\link{fur}}
}
