% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_based_iteration.R
\name{seed_based_iteration}
\alias{seed_based_iteration}
\title{Seed-Based Iteration}
\usage{
seed_based_iteration(eta, omega, input_rkgs)
}
\arguments{
\item{eta}{a subiteration length for \emph{Subiterative Convergence}, which
determines how many objects to consider in the smaller subset of objects.
The subiteration length may range between 2 and the total number of objects.}

\item{omega}{a non-negative integer for the number of repetitions of perturbing
the seed ranking. An omega value of 1 corresponds to the immediate convergence
of the \emph{Subiterative Convergence} algorithm, where \emph{Subiterative Convergence}
completes only one iteration.}

\item{input_rkgs}{a \code{n} by \code{k} matrix of \code{k} rankings of \code{n}
objects, where each row is a complete ranking.}
}
\value{
A matrix containing the consensus ranking and total Kemeny distance.
}
\description{
\emph{Seed-Based Iteration} is used in \emph{SIgFUR} to iteratively
perturb the seed ranking to improve the consensus ranking solution. The algorithm
begins by applying \emph{Subiterative Convergence} with a seed ranking of the
mean seed ranking of the input rankings. If additional repetitions are required,
then the algorithm perturbs the seed ranking by flipping its halves and applying
\emph{Subiterative Convergence} to get a new input ranking.
}
\seealso{
\code{\link{sigfur}}, \code{\link{subit_convergence}}, \code{\link{mean_seed}}
}
