% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subit_convergence.R
\name{subit_convergence}
\alias{subit_convergence}
\title{Subiterative Convergence}
\usage{
subit_convergence(eta, seed_rkg, input_rkgs, universe_rkgs = c())
}
\arguments{
\item{eta}{a subiteration length of how many objects to consider in the smaller
subset.The subiteration length may range between 2 and the total number of objects
in the ranking, and in general, a smaller \code{eta} value will take less computation time.}

\item{seed_rkg}{a initial ranking to start with. An ideal seed ranking for
Subiterative Convergence is the mean seed ranking of the input rankings.}

\item{input_rkgs}{a \code{n} by \code{k} matrix of \code{k} rankings of \code{n}
objects, where each row is a complete ranking.}

\item{universe_rkgs}{a matrix containing all possible permutations of ranking
n objects. Each row in this matrix represents one permuted ranking.}
}
\value{
A matrix containing the consensus ranking, total Kemeny distance, and
average tau correlation coefficient.
}
\description{
Perform the \emph{Subiterative Convergence} algorithm on the given seed
ranking. \emph{Subiterative Convergence} finds the consensus ranking by applying the
\emph{Modified Kemeny} algorithm on a subset of objects until the input of the iterations
converge to the same output consensus ranking.
}
\examples{
## Four input rankings of five objects
eta <- 3
seed_rkg <- c(1, 2, 3, 4, 5)
input_rkgs <- matrix(c(3, 2, 5, 1, 2, 3, 1, 2, 5, 1, 3, 4, 4, 5, 4, 5, 1, 4, 2, 3), ncol = 5)
subit_convergence(eta, seed_rkg, input_rkgs) # Determined the consensus ranking, total Kemeny
                                             # distance, and average tau correlation coefficient

## Example with eta=1
eta <- 1
seed_rkg <- c(1, 2, 3, 4, 5)
input_rkgs <- matrix(c(3, 2, 5, 1, 2, 3, 1, 2, 5, 1, 3, 4, 4, 5, 4, 5, 1, 4, 2, 3), ncol = 5)
subit_convergence(eta, seed_rkg, input_rkgs) # Shows a warning and returns seed ranking

## Included dataset of 15 input rankings of 50 objects
data(data50x15)
input_rkgs <- t(as.matrix(data50x15[, -1]))
mean_seed_rkg <- mean_seed(input_rkgs) # Use the mean seed ranking as the seed ranking
eta <- 2
subit_convergence(eta, seed_rkg = mean_seed_rkg, input_rkgs)

}
\references{
Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative
convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210.
\url{https://doi.org/10.1016/j.cor.2018.06.007}
}
\seealso{
\code{\link{mod_kemeny}}
}
