% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfur.R
\name{sigfur}
\alias{sigfur}
\title{SIgFUR}
\usage{
sigfur(
  input_rkgs,
  subit_len_list_sbi,
  omega_sbi,
  subit_len_list_fur,
  search_radius
)
}
\arguments{
\item{input_rkgs}{a \code{n} by \code{k} matrix of \code{k} rankings of \code{n}
objects, where each column is a complete ranking.}

\item{subit_len_list_sbi}{a vector containing positive integer(s) for the subiteration
lengths to \emph{Seed-Based Iteration}. Recommended values are between 2 and 8.
Smaller subiteration lengths result in shorter run-time.}

\item{omega_sbi}{a positive integer for the number of repetitions of perturbing
the seed ranking in \emph{Seed-Based Iteration}. An \code{omega_sbi} value of 1
corresponds to a single application of \emph{Subiterative Convergence}.}

\item{subit_len_list_fur}{a vector containing positive integer(s) for the subiteration
lengths to \emph{FUR}.}

\item{search_radius}{a positive integer for the maximum change in the rank of each
object in the \emph{Greedy Algorithm} and \emph{FUR}. The default value
of \code{0} considers all possible rank changes for each object. It is
recommended to use a search radius of less than or equal to \eqn{\min(30, \lfloor \mbox{n}/2 \rfloor)}{min(30, floor(n/2))}.}
}
\value{
A list containing the consensus ranking (expressed as ordering), total Kemeny distance, and average
tau correlation coefficient corresponding to the consensus ranking.
}
\description{
\emph{SIgFUR} applies \emph{Seed-Based Iteration}, \emph{Greedy Algorithm},
and \emph{FUR} in sequence for each element of \code{subit_len_list_sbi}. The
\emph{mean seed ranking} is used as the input to \emph{Seed-Based Iteration}.
The best of all output rankings from \emph{FUR} is considered as the consensus
ranking.
}
\examples{
## Four input rankings of five objects
input_rkgs <- matrix(c(3, 2, 5, 4, 1, 2, 3, 1, 5, 4, 5, 1, 3, 4, 2, 1, 2, 4, 5, 3),
    byrow = FALSE, ncol = 4)
subit_len_list_sbi <- c(2:3)
omega_sbi <- 10
subit_len_list_fur <- c(2:3)
search_radius <- 1
sigfur(input_rkgs, subit_len_list_sbi, omega_sbi, subit_len_list_fur, search_radius)
# Determined the consensus ranking, total Kemeny distance, and average tau correlation coefficient

## Included dataset of 15 input rankings of 50 objects
data(data50x15)
input_rkgs <- as.matrix(data50x15[, -1])
subit_len_list_sbi <- c(3)
omega_sbi <- 5
subit_len_list_fur <- c(2:3)
search_radius <- 1
sigfur(input_rkgs, subit_len_list_sbi, omega_sbi, subit_len_list_fur, search_radius)

}
\references{
Badal, P. S., & Das, A. (2018). Efficient algorithms using subiterative
convergence for Kemeny ranking problem. Computers & Operations Research, 98, 198-210.
\doi{10.1016/j.cor.2018.06.007}
}
\seealso{
\code{\link{seed_based_iteration}}, \code{\link{rap_greedy_alg}}, \code{\link{fur}}, \code{\link{mean_seed}}
}
